+++ accord-core/src/test/java/accord/coordinate/CoordinateTransactionTest.java
@@ -67,6 +67,7 @@ import accord.utils.async.AsyncResult;
 import static accord.Utils.id;
 import static accord.Utils.ids;
 import static accord.Utils.ranges;
+import static accord.Utils.spinUntilSuccess;
 import static accord.Utils.writeTxn;
 import static accord.impl.IntKey.key;
 import static accord.impl.IntKey.keys;
@@ -77,7 +78,6 @@ import static accord.primitives.Routable.Domain.Range;
 import static accord.primitives.Txn.Kind.Read;
 import static accord.primitives.Txn.Kind.Write;
 import static accord.utils.Invariants.checkState;
-import static accord.utils.TestUtils.*;
 import static accord.utils.async.AsyncChains.getUninterruptibly;
 import static com.google.common.base.Predicates.alwaysTrue;
 import static java.lang.Thread.sleep;
